refactor webviews
