+++ firebase-vscode/webpack.common.js
@@ -219,11 +219,11 @@ const extensionConfig = {
   },
 };
 
-function makeWebConfig(entryName) {
+function makeWebConfig(entryName, entryPath = "") {
   return {
     name: entryName,
     mode: "none", // this leaves the source code as close as possible to the original (when packaging we set this to 'production')
-    entry: `./webviews/${entryName}.entry.tsx`,
+    entry: `./webviews/${entryPath}${entryName}.entry.tsx`,
     output: {
       // the bundle is stored in the 'dist' folder (check package.json), 📖 -> https://webpack.js.org/configuration/output/
       path: path.resolve(__dirname, "dist"),
@@ -309,4 +309,9 @@ module.exports = [
     .filter((filename) => filename.match(/\.entry\.tsx/))
     .map((filename) => filename.replace(/\.entry\.tsx/, ""))
     .map((name) => makeWebConfig(name)),
+  ...fs
+    .readdirSync("webviews/data-connect")
+    .filter((filename) => filename.match(/\.entry\.tsx/))
+    .map((filename) => filename.replace(/\.entry\.tsx/, ""))
+    .map((name) => makeWebConfig(name, "data-connect/" /** entryPath */)),
 ];
