+++ firebase-vscode/webviews/data-connect/DataConnectExecutionResultsApp.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { useBroker } from "./globals/html-broker";
-import { Label } from "./components/ui/Text";
+import { useBroker } from "../globals/html-broker";
+import { Label } from "../components/ui/Text";
 import style from "./data-connect-execution-results.entry.scss";
-import { SerializedError } from "../common/error";
+import { SerializedError } from "../../common/error";
 import { ExecutionResult, GraphQLError } from "graphql";
-import { isExecutionResult } from "../common/graphql";
+import { isExecutionResult } from "../graphql";
 
 // Prevent webpack from removing the `style` import above
 style;
