+++ firebase-vscode/webviews/data-connect/data-connect-execution-configuration.entry.scss
@@ -1,4 +1,4 @@
-@import "./globals/index.scss";
+@import "../globals/index.scss";
 
 textarea {
   border-radius: 5px;
