+++ firebase-vscode/webviews/data-connect/data-connect-execution-results.entry.scss
@@ -1,4 +1,4 @@
-@import "./globals/index.scss";
+@import "../globals/index.scss";
 
 .l1 {
   text-transform: capitalize;
